+++ src/battle/battle_pokemon.c
@@ -13,8 +13,8 @@
 #include "bag.h"
 #include "battle.h"
 #include "overlay.h"
+#include "heal_after_battle.h"
 #include "pokemon.h"
-#include "save.h"
 
 #ifdef DEBUG_BATTLE_SCENARIOS
 #include "test_battle.h"
@@ -32,36 +32,6 @@ void CT_EncountSendOutMessageParamMake(void *bw, struct CLIENT_PARAM *cp, struct
 void TryRevertFormChange(struct BattleStruct *sp, void *bw, int client_no);
 void BattleEndRevertFormChange(struct BattleSystem *bw);
 
-#ifdef HEAL_AFTER_BATTLE
-static void HealPartyMon(struct PartyPokemon *mon)
-{
-    u32 maxhp;
-    u32 status = 0;
-
-    if (!GetMonData(mon, MON_DATA_SPECIES_EXISTS, NULL)) {
-        return;
-    }
-
-    maxhp = GetMonData(mon, MON_DATA_MAXHP, NULL);
-    SetMonData(mon, MON_DATA_HP, &maxhp);
-    SetMonData(mon, MON_DATA_STATUS, &status);
-    RestoreBoxMonPP(&mon->box);
-}
-
-static void BattleEndHealParty(struct BattleSystem *bw)
-{
-    int i;
-    struct Party *party = SaveData_GetPlayerPartyPtr(SaveBlock2_get());
-
-    for (i = 0; i < party->count; i++) {
-        HealPartyMon(&party->members[i]);
-    }
-
-    for (i = 0; i < BattleWorkPokeCountGet(bw, 0); i++) {
-        HealPartyMon(BattleWorkPokemonParamGet(bw, 0, i));
-    }
-}
-#endif // HEAL_AFTER_BATTLE
 // void ClearBattleMonFlags(struct BattleStruct *sp, int client);
 // u32 GetAdjustedMoveTypeBasics(struct BattleStruct *sp, u32 move, u32 ability, u32 type);
 // u32 GetAdjustedMoveType(struct BattleStruct *sp, u32 client, u32 move);
@@ -1144,7 +1114,7 @@ void BattleEndRevertFormChange(struct BattleSystem *bw)
 #endif // RESTORE_ITEMS_AT_BATTLE_END
 
 #ifdef HEAL_AFTER_BATTLE
-    BattleEndHealParty(bw);
+    HealAfterBattle_HealParty(bw);
 #endif // HEAL_AFTER_BATTLE
 
 #ifdef DEBUG_BATTLE_SCENARIOS